# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": " API",
        "description": "Women Listed Swagger API DOCS",
        "license": {
            "name": "Women Listed"
        }
    },
    "host": "",
    "basePath": "/api",
...
    "schemes": ["https", "http"],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/category": {
            "post": {
                "tags": ["CATEGORY"],
                "summary": "Category created successfully",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Creating category",
                        "required": true,
                        "schema": {
                            "$ref": "#/request/create-category"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid Object"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Validation errors"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Category created Successfully",
                        "schema": {
                            "$ref": "#/response/create-category"
                        }
                    }
                }
            },
            "get": {
                "tags": ["CATEGORY"],
                "summary": "category record",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "header",
                        "name": "authorization",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "description": "page no",
                        "required": true
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "limit of a record..By default 20",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "key",
                        "description": "search on category name",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "_id",
                        "description": "details of a category",
                        "required": false
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid Object"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "No record found!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Category record",
                        "schema": {
                            "$ref": "#/response/create-category"
                        }
                    }
                }
            }
        },
...
    },
...
}